pyproject.toml
[tool.poetry]
name = "firsttraining"
version = "0.1.0"
description = ""
authors = ["Celine <celine4231@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.1.0"
mypy = "^1.15.0"
fastapi = {extras = ["standard"], version = "^0.115.12"}
sqlalchemy = "^2.0.41"
pydantic = "^2.11.4"
celery = {extras = ["beat"], version = "^5.5.2"}
docker = "^7.1.0"
bs4 = "^0.0.2"
requests = "^2.32.3"
pymysql = "^1.1.1"
redis = "^6.1.0"
pytest = "^8.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
